Made jupyter embed file argument optional
Storing plotly files in notebooks makes the notebooks very big. 
A simple plotly plot is 5MB. The notebooks have an autosave function. When embedding the plot a file of a few MB is written every minute when the autosave triggers. 
Therefore in my opinion the option to save to a file is necessary. 

htmlwidgets::saveWidget is no option, as this does not display the plot in the notebook.

Additionally: 
base64enc does not work for me. I don't know what it is supposed to do.